pyproject.toml
[project]
name = "rpi-sonsor-client"
version = "0.1.0"
description = "Sensor client for Raspberry Pi"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adafruit-circuitpython-scd4x>=1.4.8",
    "pyyaml>=6.0.2",
    "redis>=6.2.0",
    "rpi-gpio>=0.7.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.7",
]
